+++ modulo5/Projeto-cookenu/src/index.ts
@@ -1,12 +1,5 @@
 import { app } from "./app";
-import { MysqlUsersRepository } from "./repositories/implementations/MysqlUsersRepository";
-import { CreateUserController } from "./useCases/CreateUser/CreateUserCrontroller";
+import { CreateUserController } from "./endPoints/createUser";
 
-import { CreateUserUseCase } from "./useCases/CreateUser/CreateUserUseCase";
-
-const mysqlUsersRepository = new MysqlUsersRepository();
-
-const createUserUseCase = new CreateUserUseCase(mysqlUsersRepository);
-
-const createUserController = new CreateUserController(createUserUseCase);
-app.post("/create/user", createUserController.handle);
+const createUserController = new CreateUserController();
+app.post("/create/user", createUserController.createUser);
