+++ modulo5/Projeto-cookenu/src/repositories/IUserRepositories.ts
@@ -2,5 +2,10 @@ import { User } from "../entities/User";
 
 export interface IUserRepositories {
   findByEmail(email: string): Promise<User>;
-  save(user: User): Promise<void>;
+  save(
+    id: string,
+    name: string,
+    email: string,
+    password: string
+  ): Promise<void>;
 }
