--- modulo5/Projeto-cookenu/src/useCases/CreateUser/CreateUserDTO.ts
@@ -1,5 +1,0 @@
-export interface ICreateUserRequestDto {
-  name: string;
-  email: string;
-  password: string;
-}
